+++ tests/unit/transforms/test_webpage_transform.py
@@ -1,3 +1,4 @@
+from unittest.mock import Mock
 from autolabel.transforms.webpage_transform import WebpageTransform
 import pytest
 
@@ -7,12 +8,16 @@ pytest_plugins = ("pytest_asyncio",)
 @pytest.mark.asyncio
 async def test_webpage_transform():
     # Initialize the transform class
+    webpage_transform_mock = Mock(spec=WebpageTransform)
+    webpage_transform_mock.apply.return_value = {
+        "webpage_content": "test_content",
+    }
     transform = WebpageTransform(
         output_columns={
             "content_column": "webpage_content",
-            "metadata_column": "metadata",
         },
         url_column="url",
+        scrapingbee_api_key="test_key",
         cache=None,
     )
 
@@ -21,9 +26,10 @@ async def test_webpage_transform():
     # Transform the row
     transformed_row = await transform.apply(row)
     # Check the output
-    assert set(transformed_row.keys()) == set(["webpage_content", "metadata"])
+    assert set(transformed_row.keys()) == set(
+        ["webpage_content", "webpage_transform_error"]
+    )
     assert isinstance(transformed_row["webpage_content"], str)
-    assert isinstance(transformed_row["metadata"], dict)
     assert len(transformed_row["webpage_content"]) > 0
 
 
@@ -35,32 +41,7 @@ async def test_error_handling():
             "content_column": "webpage_content",
         },
         url_column="url",
-        cache=None,
-    )
-
-    # Create a mock row
-    row = {"url": "bad_url"}
-    # Transform the row
-    transformed_row = await transform.apply(row)
-    # Check the output
-    assert set(transformed_row.keys()) == set(
-        ["webpage_content", "webpage_transform_error"]
-    )
-    assert transformed_row["webpage_content"] == "NO_TRANSFORM"
-    assert (
-        transformed_row["webpage_transform_error"]
-        == "Request URL is missing an 'http://' or 'https://' protocol."
-    )
-
-
-@pytest.mark.asyncio
-async def test_empty_url():
-    # Initialize the transform class
-    transform = WebpageTransform(
-        output_columns={
-            "content_column": "webpage_content",
-        },
-        url_column="url",
+        scrapingbee_api_key="test_key",
         cache=None,
     )
 
@@ -76,30 +57,4 @@ async def test_empty_url():
     assert (
         transformed_row["webpage_transform_error"]
         == "INVALID_INPUT: Empty url in row {'url': 'NO_TRANSFORM'}"
-    )
-
-
-@pytest.mark.asyncio
-async def test_unreachable_url():
-    # Initialize the transform class
-    transform = WebpageTransform(
-        output_columns={
-            "content_column": "webpage_content",
-        },
-        url_column="url",
-        cache=None,
-    )
-
-    # Create a mock row
-    row = {"url": "http://portal.net.kp/"}
-    # Transform the row
-    transformed_row = await transform.apply(row)
-    # Check the output
-    assert set(transformed_row.keys()) == set(
-        ["webpage_content", "webpage_scrape_error"]
-    )
-    assert transformed_row["webpage_content"] == "NO_TRANSFORM"
-    assert (
-        transformed_row["webpage_scrape_error"]
-        == "TRANSFORM_TIMEOUT: Timeout when fetching content from URL"
     )
